Add darkmode colors for the catalog tab
Added dark mode versions for the tab colors in colors.xml
into night/colors.xml, because there was no dark variation
before. This is the tab shown on catalog page.
Placed all the colors used in the tabs into the same
spot in the files, because before the colors in use were
from random spots in colors.xml. Changed the color names
to match the use cases. Deleted one unused color.

Ref EKIR-94