Fix button styles
Added a EkirjastoGreen stroke to the
contained buttons and adjusted the button
color to lightest green. In outlined buttons
added the correct outline color so it's green
for both dark and light mode. Also changed the
colorPrimary to EkirjastoWhite for the outline
buttons. Without thatthere was some light
green tint in the secondary buttons.

Ref SIMPLYE-297