Search for build_libraries.toml from multiple paths
The ekirjasto-android-platform submodule was removed from this
repository, in order to make it easier to manage dependencies in
ekirjasto-android-core. This broke building this repository as a
standalone module.

This commit adds code to search for build_libraries.toml from multiple
different paths, so that this project can be built as a standalone
module when it has been cloned as a part of ekirjasto-android-core, or
optionally by manually cloning the ekirjasto-android-platform repo.