Increase touch target size of back button
Increased the touch target size and moved the
button to the edge of the view. Since the arrow
grew in size after scaling, added padding to
scale the button back down. Padding also affected
the ekirjasto logo on the toolbar so set the padding
to zero when it's the actual logo.

Ref SIMPLYE-371