Ensure logo spot always shows when needed
IconLogoLast is forgotten when MainApplication recreate is run,
so in order for the logo to show on account page when we return
after changing textsize, we set the logo as the
ekirjasto logo when toolbar recinfigure call is done, if
there is no iconLogoLast. This is fine as the logo is always the
same.

Ref EKIR-297